update : inputview outputivew로 나눔
