+++ src/main/kotlin/baseball/view/OutputView.kt
@@ -1,11 +1,7 @@
 package baseball.view
 
-class BaseBallView {
-    fun printPlayGame() = println(START_MESSAGE)
-    fun printInputUser() = print(INPUT_USER_MESSAGE)
+class OutputView {
     fun printGameOver() = println(GAME_OVER_MESSAGE)
-    fun printContinue() = println(CHECK_CONTINUE_MESSAGE)
-
 
     fun printBaseBallResult(strike: Int, ball: Int) {
         if (ball == 0 && strike == 0) {
@@ -21,12 +17,9 @@ class BaseBallView {
     }
 
     companion object {
-        const val START_MESSAGE = "숫자 야구 게임을 시작합니다."
-        const val INPUT_USER_MESSAGE = "숫자를 입력해주세요 : "
         const val STRIKE_MESSAGE = "스트라이크"
         const val BALL_MESSAGE = "볼"
         const val NOTING_MESSAGE = "낫싱"
         const val GAME_OVER_MESSAGE = "3개의 숫자를 모두 맞히셨습니다! 게임 종료"
-        const val CHECK_CONTINUE_MESSAGE = "게임을 새로 시작하려면 1, 종료하려면 2를 입력하세요."
     }
 }
